web/requirements.txt
flask
flask_restful
pymongo
bcrypt
numpy
requests
tensorflow
keras
pillow
flask-cors
